feat: Completed Auth system and Payments integration with Stripe

+++ actions/block.ts
@@ -1,5 +1,6 @@
 "use server";
-import { getSelf } from "@/lib/auth-service";
+
+import { currentUser } from "@/lib/auth";
 import { blockUser, unblockUser } from "@/lib/block-service";
 import { RoomServiceClient } from "livekit-server-sdk";
 import { revalidatePath } from "next/cache";
@@ -11,7 +12,10 @@ const roomService = new RoomServiceClient(
 );
 
 export const onBlock = async (id: string) => {
-  const self = await getSelf();
+  const self = await currentUser();
+  if (!self) {
+    throw new Error("User not found");
+  }
 
   let blockedUser;
   try {
@@ -21,18 +25,27 @@ export const onBlock = async (id: string) => {
   }
 
   try {
-    await roomService.removeParticipant(self.id, id);
+    if (self.id) {
+      await roomService.removeParticipant(self.id, id);
+    } else {
+      throw new Error("User ID is undefined");
+    }
   } catch {
     // this means user is not in the room
   }
 
-  revalidatePath(`/u/${self.username}/community`);
+  if (self) {
+    revalidatePath(`/u/${self.username}/community`);
+  }
 
   return blockedUser;
 };
 
 export const onUnblock = async (id: string) => {
-  const self = await getSelf();
+  const self = await currentUser();
+  if (!self) {
+    throw new Error("User not found");
+  }
   const unblockedUser = await unblockUser(id);
 
   revalidatePath(`/u/${self.username}/community`);
